+++ AutoLedger/AutoLedger/App/AutoLedgerApp.swift
@@ -41,6 +41,11 @@ struct AutoLedgerApp: App {
                 AutoLedgerRootView()
             }
         }
+        #if targetEnvironment(macCatalyst)
+        .commands {
+            AutoLedgerMacCommands()
+        }
+        #endif
     }
 }